requirements.txt
mcp>=1.1.0
qdrant-client>=1.9.0
sentence-transformers>=2.6.1
python-dotenv>=1.0.1
requests>=2.32.3
